requirements.txt
django
django-chunked-upload